add: spring bone json scheme
